requirements.txt
Flask==2.3.2
azure-ai-documentintelligence==1.0.0
openai==1.3.0